fix: rename sharedStyles to Layout.Styles and move to styled folder
