+++ src/app/components/Header.tsx
@@ -1,16 +1,11 @@
 import Image from "next/image";
-import * as S from "./sharedStyles";
+import * as S from "./styled/Layout.styled";
 
 export default function Header() {
   return (
     <S.Header>
       <div style={{ display: "flex", flexDirection: "row", gap: "1.25rem" }}>
-        <Image
-          alt="evob-logo"
-          src="/icons/menu.svg"
-          width={20}
-          height={20}
-        />
+        <Image alt="evob-logo" src="/icons/menu.svg" width={20} height={20} />
         <S.Button aria-label="botao lupa">
           <Image
             alt="evob-logo"
@@ -24,7 +19,6 @@ export default function Header() {
       <S.Button aria-label="botao login">
         <Image alt="evob-logo" src="/icons/login.svg" width={32} height={32} />
       </S.Button>
-
     </S.Header>
   );
 }
